rename: edit 폴더 components 폴더에서 utils 폴더로 이동
